Remove $loopback-address, $local-host, and $local-host-name Add function local-host-name
These were no constant values, but variables. Additionally, they
confused me multiple times, since initialized by start-sockets(),
which used gethostname and passed that to gethostbyname - which
does not work on poorly configured computers which do not have DNS
or /etc/hosts setup properly.

This also updates all clients in the opendylan repository and updates
the documentation to reflect the changes.