Add a platforms: keyword to each non-generic .lid file.
Generated from the contents of sources/registry. This makes it
possible for the `dylan-tool` library to correctly generate registry
entries only for .lid files that work on the current platform.

This isn't ideal, in that it redundantly stores platform information
and we will need to remember to add this keyword to new, non-generic
.lid files as they're created. Any other solution I can think of
requires a change in the LID spec, which is a bigger job than I want
to take on right now.